code/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
xhtml2pdf>=0.2.11
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
